+++ src/bops-functions/prebuilt-functions/system/get-system-function.ts
@@ -19,7 +19,7 @@ export const getSystemFunctionFunctionInformation : InternalMetaFunction = {
     modulePackage: { type: "string", required: false },
     moduleType: { type: "string", required: true },
   },
-    output: {
+  output: {
     callableFunction: { type: "function", required: false },
     found: { type: "boolean", required: true },
   },